# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Single sign-on (сервис единого входа)",
        "title": "SSO",
        "contact": {
            "name": "Alnovi",
            "url": "https://github.com/alnovi"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/alnovi/sso/refs/heads/master/LICENSE.md"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8081",
    "paths": {
        "/admin": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Раздел администратора. Управление клиентами, пользователями и т.д.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Раздел администратора",
                "operationId": "AdminHome",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/admin/callback": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Аутентификация пользователя по коду. Вызывается автоматически после ввода логина и пароля.",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Аутентификация пользователя",
                "operationId": "AdminCallback",
                "responses": {
                    "302": {
                        "description": "Found",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/admin"
                            }
                        }
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Выход из раздела администратора",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Выход из системы",
                "operationId": "AdminLogout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Список клиентов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Список клиентов",
                "operationId": "ClientsList",
                "responses": {
                    "200": {
                        "description": "Информация о клиентах",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.Client"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Добавление нового клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Добавление клиенте",
                "operationId": "ClientsCreate",
                "parameters": [
                    {
                        "description": "Данные клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/response.Client"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/clients/{id}": {
            "get": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Информация о клиенте",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Клиент",
                "operationId": "ClientsGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор клиента",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/response.Client"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Изменение клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Изменение клиенте",
                "operationId": "ClientsUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор клиента",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateClient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/response.Client"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Удаление клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Удаление клиенте",
                "operationId": "ClientsDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор клиента",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/response.Client"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/clients/{id}/restore": {
            "post": {
                "security": [
                    {
                        "JWT-Access": []
                    }
                ],
                "description": "Восстановление удаленного клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Api Clients"
                ],
                "summary": "Восстановление клиента",
                "operationId": "ClientsRestore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор клиента",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о клиенте",
                        "schema": {
                            "$ref": "#/definitions/response.Client"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateClient": {
            "type": "object",
            "required": [
                "callback",
                "id",
                "name"
            ],
            "properties": {
                "callback": {
                    "type": "string",
                    "maxLength": 250
                },
                "icon": {
                    "type": "string",
                    "maxLength": 250
                },
                "id": {
                    "type": "string",
                    "maxLength": 30,
                    "minLength": 3
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5
                },
                "secret": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                }
            }
        },
...
        "request.UpdateClient": {
            "type": "object",
            "required": [
                "callback",
                "name",
                "secret"
            ],
            "properties": {
                "callback": {
                    "type": "string",
                    "maxLength": 250
                },
                "icon": {
                    "type": "string",
                    "maxLength": 250
                },
                "name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 5
                },
                "secret": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                }
            }
        },
...
        "response.Client": {
            "type": "object",
            "properties": {
                "callback": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_system": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT-Access": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}